refactor(auth): Extract token decoding logic to a helper func
Decoding of a token to a convenient Auth struct is moved to a separate
function to make the code a bit cleaner. Also it now uses a bit more
functional approach instead of match statements.

And finally it also bumps `frank_jwt` dependency from 3.1 to 3.1.1 as
author decided to break API in minor update: decode now takes &str
instead of &String.